LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/svg-intrinsic-size-008-expected.txt

No intrinsic attributes:
viewBox and width:
viewBox and height:
viewBox, width, height:
Just viewBox:

PASS svg 1
PASS svg 2
PASS svg 3
PASS svg 4
FAIL svg 5 assert_equals:
<svg viewBox="0 0 4 1" data-expected-client-width="400" data-expected-client-height="100"></svg>
clientWidth expected 400 but got 300
